Don't unpack RPK data directly into GROMs
The compressed data may contain backreferences into the "garbage"
part of the GROMs, which will fail if that part of the original dump
does not match the garbage generated by the GROM logic.

Unpacking into HyperRAM (allowing the "garbage" parts to be anything)
and then copying the data into GROM fixes the issue.